convert esbuild config to typescript
