+++ apps/app-esbuild/tsconfig.json
@@ -26,6 +26,9 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true
   },
-  "include": ["src"],
-  "exclude": ["node_modules", "dist", "dev"]
+
+  "include": ["src", "../../@types"],
+  "exclude": ["node_modules", "dist", "dev"],
+
+  "references": [{ "path": "./tsconfig.node.json" }]
 }
